% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{build_api_url}
\alias{build_api_url}
\title{Build a Zenodo records API URL}
\usage{
build_api_url(query, community, size, page)
}
\arguments{
\item{query}{Optional search query string.}

\item{community}{Optional community id.}

\item{size}{Page size for the request.}

\item{page}{Page number for the request.}
}
\value{
A fully-qualified URL string.
}
\description{
Build a Zenodo records API URL
}
