% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{build_concept_map}
\alias{build_concept_map}
\title{Map concept ids to the latest version record id}
\usage{
build_concept_map(records)
}
\arguments{
\item{records}{List of Zenodo records.}
}
\value{
Named list mapping concept id to record id.
}
\description{
Map concept ids to the latest version record id
}
