% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{build_query}
\alias{build_query}
\title{Build a Zenodo query string}
\usage{
build_query(query_text)
}
\arguments{
\item{query_text}{User-provided query.}
}
\value{
Query string or NULL.
}
\description{
Build a Zenodo query string
}
