% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{extract_doi}
\alias{extract_doi}
\title{Extract a DOI from identifier text}
\usage{
extract_doi(identifier)
}
\arguments{
\item{identifier}{Identifier string.}
}
\value{
DOI string or empty string.
}
\description{
Extract a DOI from identifier text
}
