% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{extract_relations}
\alias{extract_relations}
\title{Extract relation types present between community records}
\usage{
extract_relations(records, community_ids = NULL, concept_map = NULL)
}
\arguments{
\item{records}{List of Zenodo records.}

\item{community_ids}{Character vector of community record ids.}

\item{concept_map}{Optional map of concept id to record id.}
}
\value{
Character vector of relation labels.
}
\description{
Extract relation types present between community records
}
