% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{filter_by_keywords}
\alias{filter_by_keywords}
\title{Filter records by keyword tokens}
\usage{
filter_by_keywords(records, keywords_text)
}
\arguments{
\item{records}{List of Zenodo records.}

\item{keywords_text}{Vector or string of keywords.}
}
\value{
Filtered list of records.
}
\description{
Filter records by keyword tokens
}
