% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{is_valid_record}
\alias{is_valid_record}
\title{Validate a record object}
\usage{
is_valid_record(rec)
}
\arguments{
\item{rec}{Record candidate.}
}
\value{
Logical scalar.
}
\description{
Validate a record object
}
