% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{is_zenodo_doi}
\alias{is_zenodo_doi}
\title{Check if an identifier is a Zenodo DOI}
\usage{
is_zenodo_doi(identifier)
}
\arguments{
\item{identifier}{Identifier string.}
}
\value{
Logical scalar.
}
\description{
Check if an identifier is a Zenodo DOI
}
