% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{normalize_keywords}
\alias{normalize_keywords}
\title{Normalize keywords for matching}
\usage{
normalize_keywords(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Normalized character vector.
}
\description{
Normalize keywords for matching
}
