% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{normalize_records}
\alias{normalize_records}
\title{Normalize records to a list of record objects}
\usage{
normalize_records(records)
}
\arguments{
\item{records}{List or data.frame.}
}
\value{
List of records.
}
\description{
Normalize records to a list of record objects
}
