% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{records_to_table}
\alias{records_to_table}
\title{Build a metadata table for display}
\usage{
records_to_table(records)
}
\arguments{
\item{records}{List of Zenodo records.}
}
\value{
Data frame with display columns.
}
\description{
Build a metadata table for display
}
