% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{sanitize_records}
\alias{sanitize_records}
\title{Sanitize records by normalizing and filtering invalid entries}
\usage{
sanitize_records(records)
}
\arguments{
\item{records}{List or data.frame.}
}
\value{
List of valid records.
}
\description{
Sanitize records by normalizing and filtering invalid entries
}
