% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny server logic for the Zenodo network explorer}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Shiny input.}

\item{output}{Shiny output.}

\item{session}{Shiny session.}
}
\description{
Shiny server logic for the Zenodo network explorer
}
