% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{Shiny UI for the Zenodo network explorer}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
A Shiny UI object.
}
\description{
Shiny UI for the Zenodo network explorer
}
\keyword{datasets}
