% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{zenodo_id_from_identifier}
\alias{zenodo_id_from_identifier}
\title{Extract a Zenodo record id from DOI or record URL}
\usage{
zenodo_id_from_identifier(identifier)
}
\arguments{
\item{identifier}{Identifier string.}
}
\value{
Record id as character or NA if not found.
}
\description{
Extract a Zenodo record id from DOI or record URL
}
