% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zenodo_search_auth}
\alias{zenodo_search_auth}
\title{Search Zenodo records (authenticated)}
\usage{
zenodo_search_auth(query, community = NULL, size = 200, page = 1, token = NULL)
}
\arguments{
\item{query}{Optional search query string.}

\item{community}{Optional community id.}

\item{size}{Page size for the request.}

\item{page}{Page number for the request.}

\item{token}{Zenodo API token.}
}
\value{
A list parsed from the Zenodo JSON response.
}
\description{
Search Zenodo records (authenticated)
}
